Add fetch history feature and docs
